\name{stats301}
\alias{stats301}
\title{Function to determine 301 stats from an input data.frame}
\usage{
stats301(myDF)
}
\arguments{
  \item{myDF}{input dataframe of 301 stats}
}
\description{
This takes an input dataframe of a specified format and
calculates several 301 stats.  Output file is a data.frame.
}

