\name{winPerc301}
\alias{winPerc301}
\title{Function to determine 301 winning percent from an input data.frame}
\usage{
winPerc301(myDF)
}
\arguments{
  \item{myDF}{input dataframe of 301 stats}
}
\description{
This takes an input dataframe of a specified format and
caluclates the winning percentage
}

