\name{winPercCricket}
\alias{winPercCricket}
\title{Function to determine crickte winning percent from an input data.frame}
\usage{
winPercCricket(myDF)
}
\arguments{
  \item{myDF}{input dataframe of cricket stats}
}
\description{
This takes an input dataframe of a specified format and
caluclates the winning percentage
}

